BILL-297: Move superagent declaration out of src into typings
